feat: Add env built-in
Remove set_env.c and added its functions to env.c
Create env built-in to print env variable
