+++ libs/libft/Makefile
@@ -6,7 +6,7 @@
 #    By: vcastilh <vcastilh@student.42sp.org.br>    +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2021/08/18 10:20:32 by vcastilh          #+#    #+#              #
-#    Updated: 2022/10/10 20:43:07 by coder            ###   ########.fr        #
+#    Updated: 2022/10/18 22:24:21 by vcastilh         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -54,6 +54,8 @@ FILES = ft_isalpha.c \
 		ft_lstsize.c \
 		ft_lstlast.c \
 		ft_lstadd_back.c \
+		ft_lstdelone.c \
+		ft_lstclear.c
 
 
 OBJFILES = $(FILES:%.c=%.o)
