+++ libs/libft/ft_lstnew.c
@@ -6,7 +6,7 @@
 /*   By: coder <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/15 14:28:57 by coder             #+#    #+#             */
-/*   Updated: 2022/10/10 20:31:11 by coder            ###   ########.fr       */
+/*   Updated: 2022/10/19 04:16:19 by coder            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,9 @@ t_list	*ft_lstnew(void *content)
 	lst = malloc(sizeof(t_list));
 	if (lst == NULL)
 		return (NULL);
-	lst->var = content;
+	lst->var = ft_strdup(content);
+	if (lst->var == NULL)
+		return (NULL);
 	lst->next = NULL;
 	return (lst);
 }
