+++ libs/libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: vcastilh <vcastilh@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/23 08:47:36 by vcastilh          #+#    #+#             */
-/*   Updated: 2022/10/10 20:30:41 by coder            ###   ########.fr       */
+/*   Updated: 2022/10/18 22:23:33 by vcastilh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -69,7 +69,7 @@ t_list		*ft_lstnew(void	*content);
 //int			ft_lstsize(t_list *lst);
 t_list		*ft_lstlast(t_list *lst);
 void		ft_lstadd_back(t_list **lst, t_list *new_node);
-//void		ft_lstdelone(t_list *lst, void (*del)(void *));
-//void		ft_lstclear(t_list **lst, void (*del)(void *));
+void		ft_lstdelone(t_list *lst, void (*del)(void *));
+void		ft_lstclear(t_list **lst, void (*del)(void *));
 //void		ft_lstiter(t_list *lst, void (*)(void *));
 #endif
